+++ packages/core/mashroom-utils/src/ResourceFetchError.ts
@@ -1,0 +1,11 @@
+import {mergeStackTrace} from './error-utils';
+
+export default class ResourceFetchError extends Error {
+    constructor(message: string, rootCause?: Error) {
+        super(message);
+        this.name = 'ResourceFetchError';
+        if (rootCause) {
+            this.stack = mergeStackTrace(this.stack ?? '', rootCause.stack);
+        }
+    }
+}