+++ packages/core/mashroom-utils/src/ResourceNotFoundError.ts
@@ -1,0 +1,7 @@
+
+export default class ResourceNotFoundError extends Error {
+    constructor(message?: string) {
+        super(message);
+        this.name = 'ResourceNotFoundError';
+    }
+}