+++ packages/core/mashroom-utils/src/ResourceTypeNotSupportedError.ts
@@ -1,0 +1,7 @@
+
+export default class ResourceTypeNotSupportedError extends Error {
+    constructor(message?: string) {
+        super(message);
+        this.name = 'UnsupportedResourceTypeError';
+    }
+}