+++ packages/plugin-packages/mashroom-portal/src/backend/dev-server.ts
@@ -32,6 +32,11 @@ setPortalPluginConfig({
     autoExtendAuthentication: false,
     ignoreMissingAppsOnPages: false,
     versionHashSalt: null,
+    resourceFetchConfig: {
+        fetchTimeoutMs: 3000,
+        httpMaxSocketsPerHost: 10,
+        httpRejectUnauthorized: true,
+    },
     defaultProxyConfig: {},
     ssrConfig: {
         ssrEnable: true,