+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortaWebSocketProxyController.test.ts
@@ -13,6 +13,11 @@ setPortalPluginConfig({
     autoExtendAuthentication: false,
     ignoreMissingAppsOnPages: false,
     versionHashSalt: null,
+    resourceFetchConfig: {
+        fetchTimeoutMs: 3000,
+        httpMaxSocketsPerHost: 10,
+        httpRejectUnauthorized: true,
+    },
     defaultProxyConfig: {
         sendPermissionsHeader: false,
         restrictToRoles: [],